changed variables data types and changed stored procedures in controllers

+++ TrafficManagementApi/Models/ResultCrossroad.cs
@@ -8,14 +8,9 @@ namespace TrafficManagementApi.Models
     public class ResultCrossroad: BaseResponse
     {
         public int Id { get; set; }
-
-        public float Id_Crossroad { get; set; }
-
+        public decimal Id_Crossroad { get; set; }
         public int Id_Priority { get; set; }
-
         public DateTime Date { get; set; }
-
-        
 
     }
 }
